% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_if.R
\name{browser_if}
\alias{browser_if}
\title{Call browser if given condition is TRUE}
\usage{
browser_if(truth = stop("truth not supplied"))
}
\arguments{
\item{truth}{An epression that evaluates to a logical value}
}
\value{
Nothing
}
\description{
I often found myself wrapping a call to \code{browser} in an if
statement.
}
\details{
This function takes a logical expression and, if it evaluates to true,
calls \code{browser}.

This is defined as a macro so that when \code{browser} is invoked, it
have the context of the place where \code{browser_if} was called
}
