% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_tap.R
\name{btap}
\alias{btap}
\title{Call \code{browser} with \code{thing} set to the \code{tap}'d object so
it can be inspected}
\usage{
btap(thing)
}
\arguments{
\item{thing}{The object to inspect during a call to \code{browser}}
}
\value{
The object passed in to allow for method chaining
}
\description{
Call \code{browser} with \code{thing} set to the \code{tap}'d object so
it can be inspected
}
