% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{builtins2}
\alias{builtins2}
\alias{special_fns}
\alias{builtin_fns}
\title{Extension of \code{base::builtins()}.}
\usage{
builtins2(type = c("all", "closure", "builtin", "sprcial"), internal = FALSE)

special_fns(internal = FALSE)

builtin_fns(internal = FALSE)
}
\arguments{
\item{type}{a string. One of \code{"all"}, \code{"closure"}, \code{"builtin"}, or \code{"sprcial"}. \code{"all"} is equivalent to \code{base::builtins()}. \code{"closure"} retains only the functions with type \code{"closure"}, \code{"builtin"} the functions with type \code{"builtin"}, and \code{"special"} the functions with type \code{"special"}.}

\item{internal}{a logical indicating whether only \sQuote{internal}
    functions (which can be called via \code{\link[base]{.Internal}}) should
    be returned.}
}
\value{
a list of functions.
}
\description{
Extension of \code{base::builtins()}.
}
