% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cache.dated}
\alias{cache.dated}
\title{Saves an object to the cache/ dir adding the date when the object was saved}
\usage{
cache.dated(object)
}
\description{
Saves an object to the cache/ dir adding the date when the object was saved
}
\examples{
cache.dated()
}
\keyword{cache}

