% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-helpers.R
\name{call_add_ns}
\alias{call_add_ns}
\title{Add Namespace to a call.}
\usage{
call_add_ns(x, .ns)
}
\arguments{
\item{x}{an R object to convert into a simple call. It is passed to \code{encall0}.}

\item{.ns}{Namespace with which to prefix \code{.fn}. Must be a string
or symbol.}
}
\description{
Add Namespace to a call.
}
