% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_replace.R
\name{call_replace}
\alias{call_replace}
\alias{call_replace_fn}
\alias{call_replace_ns}
\alias{call_add_ns}
\title{Replace call components}
\usage{
call_replace(
  x,
  ...,
  .fn = call_name(x),
  .args = call_args(x),
  .ns = call_ns(x)
)

call_replace_fn(x, .fn)

call_replace_ns(x, .ns)

call_add_ns(x, .ns)
}
\arguments{
\item{x}{an R object to convert into a simple call. It is passed to \code{encall0}.}

\item{...}{must remain empty.}

\item{.fn}{Function to call. Must be a callable object: a string,
symbol, call, or a function.}

\item{.args}{a list.}

\item{.ns}{Namespace with which to prefix \code{.fn}. Must be a string
or symbol.}
}
\value{
a simple call.
}
\description{
Replace call components
}
