% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-helpers.R
\name{call_text}
\alias{call_text}
\title{Deparse Call}
\usage{
call_text(x, ..., .simplify = FALSE)
}
\arguments{
\item{x}{a call.}

\item{...}{empty.}

\item{.simplify}{a logical.}
}
\value{
a string.
}
\description{
Deparse Call
}
