% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na.R
\name{check_na}
\alias{check_na}
\title{Sum NA values in a dataframe by column}
\usage{
check_na(x, empty_str = FALSE)
}
\arguments{
\item{x}{dataframe to interrogate.}

\item{empty_str}{logical, should empty strings be counted as well?
Defaults to false.}
}
\value{
a named int, where the names correspond to the original variable names
and the values correspond to the number of NAs found for that variable.
}
\description{
Sum NA values in a dataframe by column
}
\examples{
df <- data.frame(x = c('a','b',NA,''), y = c(1,NA,2,NA))
check_na(df)
check_na(df, empty_str = TRUE)
}
