% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean.cache.dated}
\alias{clean.cache.dated}
\title{Deletes cache object that are older than clean.older.than.days days old
Requires the same file structure as cache.dated and load.cache.dated}
\usage{
clean.cache.dated(clean.older.than.days = 7, remove = FALSE)
}
\description{
Deletes cache object that are older than clean.older.than.days days old
Requires the same file structure as cache.dated and load.cache.dated
}
\examples{
clean.cache.dated()
}
\keyword{cache}

