% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_console.R
\name{clear_console}
\alias{clear_console}
\title{Clears the console when run}
\usage{
clear_console()
}
\value{
Nothing, but clears the console
}
\description{
Found this trick on https://stackoverflow.com/a/16084793 and it works
a treat!
}
