% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfs.R
\name{combine_pdf_folder}
\alias{combine_pdf_folder}
\title{Combine all pdfs in a folder.}
\usage{
combine_pdf_folder(folder_of_pdfs, output_name, output_path = folder_of_pdfs)
}
\arguments{
\item{folder_of_pdfs}{path to folder containing pdfs}

\item{output_name}{Name of saved file}

\item{output_path}{(str; default = folder_of_pdfs) Where should the pdfs be stored}
}
\value{
Location of the outputted pdf file
}
\description{
Combine all pdfs in a folder.
}
