% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/many_functions.R
\name{count.na}
\alias{count.na}
\title{Counts the number of NAs in a data.frame}
\usage{
count.na(data)
}
\arguments{
\item{data.}{}
}
\description{
Counts the number of NAs in a data.frame
}
\keyword{counts}

