% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_var.R
\name{drop_var}
\alias{drop_var}
\title{Drop a given variable from a given dataframe}
\usage{
drop_var(x, var)
}
\value{
Returns the dataframe with the variable removed
}
\description{
Drop a given variable from a given dataframe
}
\examples{
names(drop_var(mtcars, "mpg"))
}
