% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{empty_as_na}
\alias{empty_as_na}
\title{replace empty ("") values with "NA"}
\usage{
df %>% empty_as_na()

  # or

  empty_as_na(df)
}
\arguments{
\item{df}{the dataframe or tibble to be modified}
}
\value{
df with empty values replaced with "NA"
}
\description{
"NA" are often easier to see and deal with compared to empty
  values. Further testing is needed for this function
}
\examples{
\dontrun{
require(tibble)

col1 <- seq(1, 10, 1)
col2 <- c("", 3, 4, "", 6, 7, 8, "", 10, "")
my_df <- tibble(col1 = col1, col2 = col2)
empty_as_na(df = my_df)
}
}
