% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encall.R
\name{encall}
\alias{encall}
\alias{encall0}
\alias{encall_replace}
\alias{encalls}
\title{Objects as calls}
\usage{
encall0(x)

encall_replace(x, ..., .args = NULL, .ns = NULL)

encall(x, .simplify = TRUE)

encalls(
  ...,
  .named = TRUE,
  .ignore_empty = c("trailing", "none", "all"),
  .ignore_null = c("none", "all"),
  .unquote_names = TRUE,
  .homonyms = c("keep", "first", "last", "error"),
  .check_assign = FALSE
)
}
\arguments{
\item{x}{a string, a symbol, or a call.}

\item{...}{strings or expressions to defuse.}

\item{.named}{If \code{TRUE}, unnamed inputs are automatically named
with \code{\link[rlang:as_label]{as_label()}}. This is equivalent to applying
\code{\link[rlang:exprs_auto_name]{exprs_auto_name()}} on the result. If \code{FALSE}, unnamed elements
are left as is and, if fully unnamed, the list is given minimal
names (a vector of \code{""}). If \code{NULL}, fully unnamed results are
left with \code{NULL} names.}

\item{.ignore_empty}{Whether to ignore empty arguments. Can be one
of \code{"trailing"}, \code{"none"}, \code{"all"}. If \code{"trailing"}, only the
last argument is ignored if it is empty. Named arguments are not
considered empty.}

\item{.ignore_null}{Whether to ignore unnamed null arguments. Can be
\code{"none"} or \code{"all"}.}

\item{.unquote_names}{Whether to treat \verb{:=} as \code{=}. Unlike \code{=}, the
\verb{:=} syntax supports \link[rlang:glue-operators]{names injection}.}

\item{.homonyms}{How to treat arguments with the same name. The
default, \code{"keep"}, preserves these arguments. Set \code{.homonyms} to
\code{"first"} to only keep the first occurrences, to \code{"last"} to keep
the last occurrences, and to \code{"error"} to raise an informative
error and indicate what arguments have duplicated names.}

\item{.check_assign}{Whether to check for \verb{<-} calls. When \code{TRUE} a
warning recommends users to use \code{=} if they meant to match a
function parameter or wrap the \verb{<-} call in curly braces otherwise.
This ensures assignments are explicit.}
}
\value{
namespaced call(s).
}
\description{
Objects as calls
}
