% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-helpers.R
\name{enlang}
\alias{enlang}
\alias{enlangs0}
\alias{enlangs}
\title{Objects as calls}
\usage{
enlang(x)

enlangs0(x, .named = FALSE)

enlangs(..., .named = FALSE)
}
\arguments{
\item{x}{a string, a symbol, or a call.}

\item{.named}{If \code{TRUE}, unnamed inputs are automatically named
with \code{\link[rlang:as_label]{as_label()}}. This is equivalent to applying
\code{\link[rlang:exprs_auto_name]{exprs_auto_name()}} on the result. If \code{FALSE}, unnamed elements
are left as is and, if fully unnamed, the list is given minimal
names (a vector of \code{""}). If \code{NULL}, fully unnamed results are
left with \code{NULL} names.}

\item{...}{strings or expressions to defuse.}
}
\value{
namespaced call(s).
}
\description{
Objects as calls
}
