% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/many_functions.R
\name{error.email.f}
\alias{error.email.f}
\title{Sends error email with error log attached}
\usage{
error.email.f(error_log.path, error.display.name, admin.email, error.email)
}
\description{
Sends error email with error log attached
}
\examples{
error.email.f()
}
\keyword{email}

