% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-call.R
\name{evaluate-call}
\alias{evaluate-call}
\alias{call_eval0}
\alias{call_eval}
\title{Evaluate a call or a list of calls}
\usage{
call_eval0(x, ..., data = NULL, env = caller_env())

call_eval(x, ..., data = NULL, env = caller_env())
}
\arguments{
\item{x}{a call (\code{call_eval0}) or an R object or list of objects (\code{call_eval0}) that will be passed to \code{encall(x)}.}

\item{...}{must remain empty.}

\item{data}{a data frame, or named list or vector.}

\item{env}{the environment in which to evaluate \code{x}.}
}
\description{
Evaluate a call or a list of calls
}
