% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clip.R
\name{format_clip}
\alias{format_clip}
\title{Convert backslashes to forward slashes directly from the clipboard.}
\usage{
format_clip()
}
\value{
The string copied to the clipboard is returned with backslashes
}
\description{
Convert backslashes to forward slashes directly from the clipboard.
}
