% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czech-encoding.R
\name{get_czech_platform_locale}
\alias{get_czech_platform_locale}
\title{Returns the locale needed to process czech characters}
\usage{
get_czech_platform_locale()
}
\value{
character with specific output. 1250 on win and utf-8 on linux and mac
}
\description{
this is a wrapper to solve the fact R does not support
UTF-8 encoding on Windows
}
