% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_helpers.R
\name{get_lat_center}
\alias{get_lat_center}
\title{get latitude center}
\usage{
get_lat_center(x)
}
\arguments{
\item{x}{is a vector or column of latitude values}
}
\value{
A single numeric value that is the center (average) of @param x
}
\description{
get the latitide center of a given vector or column of latitude values
}
\examples{
\dontrun{
my_latitudes <- sample(-90:90, 100, replace = TRUE)
get_lat_center(my_latitudes)
}
}
