% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_helpers.R
\name{get_long_center}
\alias{get_long_center}
\title{get longitude center}
\usage{
get_long_center(x)
}
\arguments{
\item{x}{is a vector or column of longitude values}
}
\value{
A single numeric value that is the center (average) of x
}
\description{
get the longitude center of a given vector or column of longitude values
}
\examples{
\dontrun{
my_longitudes <- sample(-180:180, 100, replace = TRUE)
get_long_center(my_longitudes)
}
}
