% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{get_missing_percent}
\alias{get_missing_percent}
\title{get missing percent within a column of a tibble or a dataframe}
\usage{
get_missing_percent(data, missing_value, ...)
}
\arguments{
\item{data}{is the dataframe or tibble to which the column of interest belongs}

\item{missing_value}{is the value to be counted as "missing", i.e "NA" or "", or c("NA", NA, "")}

\item{...}{additional arguments passed}
}
\value{
a character string of the missing percentage of miss_col
}
\description{
get the percent of observations that are missing in a column of a tibble or dataframe.
Can be mapped or future_mapped over multiple columns in a dataframe
}
\examples{
\dontrun{
require(tibble)

col1 <- seq(1, 10, 1)
col2 <- c("NA", 5, 7, NA, 41, "NA", 6, 8, "NA", NA)
my_tibble <- tibble(col1 = col1, col2 = col2)
get_missing_percent(my_tibble$col1, missing_value = "NA")
}
}
