% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/many_functions.R
\name{grepr}
\alias{grepr}
\title{Wrapper of grep that makes it more magrittr friendly}
\usage{
grepr(string.vec, pattern)
}
\description{
Wrapper of grep that makes it more magrittr friendly
}
\keyword{grep}

