% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helprs-package.R
\docType{package}
\name{helprs-package}
\alias{helprs}
\alias{helprs-package}
\title{helprs: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/f-dallolio/helprs}
  \item Report bugs at \url{https://github.com/f-dallolio/helprs/issues}
}

}
\author{
\strong{Maintainer}: Filippo DallOlio \email{fdallolio@brocku.ca}

}
\keyword{internal}
