% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_blank.R
\name{is_blank}
\alias{is_blank}
\title{Check object to see if it's emptyish/blankish}
\usage{
is_blank(thing)
}
\arguments{
\item{thing}{An object to test}
}
\value{
TRUE if object is: NA, NULL, an empty array/list, or empty string
}
\description{
Check object to see if it's emptyish/blankish
}
