% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_present.R
\name{is_present}
\alias{is_present}
\title{Check object to see if it is NOT emptyish/blankish}
\usage{
is_present(...)
}
\arguments{
\item{...}{An object to test}
}
\value{
TRUE if object is NOT: NA, NULL, an empty array/list, or empty string
}
\description{
This is literally the inverse of \code{is_blank}
}
