% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_to_str.R
\name{lang_to_str}
\alias{lang_to_str}
\title{Convert a symbolic to string}
\usage{
lang_to_str(x, simplify = TRUE)
}
\arguments{
\item{x}{a symbol or a call.}

\item{simplify}{logical.}
}
\value{
a string.
}
\description{
Convert a symbolic to string
}
