% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_debug_tap.R
\name{ldtap}
\alias{ldtap}
\title{Log an object via debug}
\usage{
ldtap(thing)
}
\arguments{
\item{thing}{The object to log to debug}
}
\value{
The object passed in to allow for method chaining
}
\description{
Log an object via debug
}
