% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_functions.R
\name{list.class}
\alias{list.class}
\title{get the first element of the class types for every element in a list}
\usage{
list.class(x)
}
\description{
get the first element of the class types for every element in a list
}
\keyword{source}

