% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.cache.dated}
\alias{load.cache.dated}
\title{load the most recent version of an object, follows the namings conventions of cache.dated()}
\usage{
load.cache.dated(object.name)
}
\description{
load the most recent version of an object, follows the namings conventions of cache.dated()
}
\examples{
load.cache.dated()
}
\keyword{cache}

