% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.dated.cache}
\alias{load.dated.cache}
\title{load the most recent version of an object, follows the namings conventions of cache.dated()}
\usage{
load.dated.cache(object.name)
}
\description{
load the most recent version of an object, follows the namings conventions of cache.dated()
}
\examples{
cache.dated()
}
\keyword{cache}

