% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{m_and_sd_report}
\alias{m_and_sd_report}
\title{Reports mean and standard devisation in APA style}
\usage{
m_and_sd_report(vec, digits = 3)
}
\arguments{
\item{vec}{vector of numbers}

\item{digits}{how many digits to round thge number for}
}
\value{

}
\description{
Reports mean and standard devisation in APA style
}
