% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron_helpers.R
\name{make_crontab}
\alias{make_crontab}
\title{create a crontab job from an rscript}
\usage{
make_crontab(.script, .frequency, .at = NULL)
}
\arguments{
\item{.script}{the path to the script to made into a crontab job}

\item{.frequency}{frequency at which to run, i.e. "daily" or "weekly" or "minutely"}

\item{.at}{the sub-frequency at which to run, i.e. for .frequency = "daily", .at could be "8AM"}
}
\value{
a console printout of the crontab that was added for the current user
}
\description{
use the 'cronR' package to create a crontab job with one function
}
\examples{
\dontrun{
make_crontab("user/my_folder/my_script.R", .frequency = "daily", .at = "8AM")
}
}
