% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{make_date_underscored}
\alias{make_date_underscored}
\title{print the date}
\usage{
make_date_underscored()
}
\value{
a character string of the date
}
\description{
print today's date like 2019_05_21
}
\examples{
\dontrun{
make_date_underscored()
}
}
