% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_path.R
\name{make_path}
\alias{make_path}
\title{Creates a directory and any necessary parent directories}
\usage{
make_path(dir_path)
}
\arguments{
\item{dir_path}{Path to the directory that should be created}
}
\value{
Path to the directory created
}
\description{
Just a simple wrapper around dir.create to avoid errors and having to
add extra arguments
}
