% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{make_time_stamp}
\alias{make_time_stamp}
\title{print a timestamp}
\usage{
make_time_stamp()
}
\value{
a character string of the timestamp
}
\description{
print a timestamp like "20190521_0800"
}
\examples{
\dontrun{
make_time_stamp()
}
}
