% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_helpers.R
\name{make_tsibble}
\alias{make_tsibble}
\title{turn a tibble or dataframe into a tsibble}
\usage{
make_tsibble(df, date_column, group_column,
period = c("day", "week", "month", "quarter", "year"))
}
\arguments{
\item{df}{the tibble or dataframe from which to make a tsibble}

\item{date_column}{the column containing a date for each observation}

\item{group_column}{the column by which to group the data for each date}

\item{period}{the lubridate::floor_date period to floor the observation dates to}
}
\value{
a tsibble with the specified time interval as the Key
}
\description{
create a 'time series tibble' from a tibble or dataframe
}
\examples{
\dontrun{
nycflights13::weather \%>\%
  make_tsibble(df = .,
               date_column = time_hour,
               group_column = origin,
               period = "week")
}
}
