% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-fn-names.R
\name{namespace-fn-names}
\alias{namespace-fn-names}
\alias{ns_fn_names}
\alias{ns_exports_names}
\alias{ns_private_names}
\title{Get names of functions in a namespace.}
\usage{
ns_fn_names(ns, exports_only = FALSE)

ns_exports_names(ns)

ns_private_names(ns)
}
\arguments{
\item{ns}{a string indicating the name of a namespace.}

\item{exports_only}{logical. If TRUE the function retrieves the names of exported functions only. If FALSE, the function retrieves the names of all functions in a namespace.}
}
\value{
a character vector.
}
\description{
Get names of functions in a namespace.
}
