% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace-get-functions.R
\name{namespace-functions}
\alias{namespace-functions}
\alias{get_fn}
\alias{get_ns_fns}
\alias{get_export_fns}
\title{Get functions from function names and namespace}
\usage{
get_fn(x, ns = NULL, ifnotfound = NULL)

get_ns_fns(ns, ..., ifnotfound = NULL, exports_only = FALSE)

get_export_fns(ns = NULL)
}
\arguments{
\item{x}{a string (\code{get_fn}) or a character vector (\code{get_fns}).}

\item{ns}{the namespace name.}

\item{ifnotfound}{the return value of \verb{get_fn*(x, *)} when x does not exist.}

\item{names_only}{logical. If \code{TRUE}, the function returns only the names of the functions.}

\item{export_only}{logical. If \code{TRUE}, the function returns only the names of the functions.}
}
\value{
a function or list of functions.
}
\description{
Get functions from function names and namespace
}
