% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs-are.R
\name{objs-are}
\alias{objs-are}
\alias{are}
\alias{arenot}
\alias{is}
\alias{isnot}
\title{Are objects...?}
\usage{
are(.x, .p)

arenot(.x, .p, .all = FALSE)

is(.x, .p)

isnot(.x, .p)
}
\arguments{
\item{.x}{a vector or a symbolic object (symbol or call).}

\item{.p}{a function returning a logical vector.}
}
\value{
The functions \code{are} and \code{arenot} return a vector of the same length as \code{.x}. The functions \code{is} and \code{is_not} return a boolean.
}
\description{
Are objects...?
}
