% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/many_functions.R
\name{rbindlist.valid}
\alias{rbindlist.valid}
\title{Wrapper for rbindlist that checks the stack every object in the input list that's a data.table}
\usage{
rbindlist.valid(list.data.tables.x)
}
\description{
Wrapper for rbindlist that checks the stack every object in the input list that's a data.table
}
\keyword{cache}

