% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/many_functions.R
\name{rm.var}
\alias{rm.var}
\title{Remove everything but functions from the global environment}
\usage{
rm.var()
}
\description{
Remove everything but functions from the global environment
}
\keyword{rm}

