% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czech-encoding.R
\name{set_czech_locale}
\alias{set_czech_locale}
\title{Sets the encoding permanently to czech. takes into consideration the platform}
\usage{
set_czech_locale(category = "LC_ALL")
}
\arguments{
\item{category}{what locale to change. Default to LC_ALL}
}
\value{
returns the original locale before the change
}
\description{
Sets the encoding permanently to czech. takes into consideration the platform
}
