% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_functions.R
\name{source.wrapper}
\alias{source.wrapper}
\title{Wraps source of script adding email sending on error and dbWriting on success}
\usage{
source.wrapper(script.path, error_log.path, admin.email, error.email)
}
\description{
Wraps source of script adding email sending on error and dbWriting on success
}
\keyword{source}

