% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-folder.R
\name{source_folder}
\alias{source_folder}
\title{Sources files from a folder to a given depth}
\usage{
source_folder(folder, depth = 0, env = .GlobalEnv)
}
\arguments{
\item{depth}{if 0, then the depth is endless}
}
\value{
evaluates in Global env
}
\description{
Sources files from a folder to a given depth
}
