% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sets.R
\name{split_sets}
\alias{split_sets}
\title{Split a dataframe into training, test and (if required) validation sets.}
\usage{
split_sets(x, props = c(0.7, 0.3), validation = FALSE)
}
\arguments{
\item{x}{A dataframe to be split into training and testing (and, if required,
validation) sets. Must have nrow >= 20.}

\item{props}{Numeric vector of proportions to split the dataframe by. Must be
length 2 or 3. The first
element is the proportion of observations to be assigned to the training set;
the second element is the proportion of observations to be assigned to the
testing set; and the optional third element is the proportion of observations
to be assigned to the validation set. Defaults to `c(0.7, 0.3)`, where
approximately 70% of observations are assigned to the training set, and 30%
of observations are assigned to the testing set.}
}
\value{
A list of dataframes of length 2 or 3, depending on whether `validation` is set to
FALSE or TRUE.
}
\description{
Split a dataframe into training, test and (if required) validation sets.
}
\examples{
split_dfs <- split(mtcars, props = c(0.6, 0.4))
}
