% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_match_any.R
\name{str_match_any}
\alias{str_match_any}
\title{Compare strings against one or more regex patterns and return TRUE
for each string that matches at least one pattern}
\usage{
str_match_any(strings, patterns)
}
\arguments{
\item{strings}{Strings to match against the patterns}

\item{patterns}{Regexes to match against each string}
}
\value{
logical representing if a string matched at least one pattern
}
\description{
Compare strings against one or more regex patterns and return TRUE
for each string that matches at least one pattern
}
