% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tap.R
\name{tap}
\alias{tap}
\title{Executes a callback function on an object before passing
the object along}
\usage{
tap(thing, callback)
}
\arguments{
\item{thing}{The object to pass to the callback and then return}

\item{callback}{The function that receives the object and does something with it}
}
\value{
The object passed in to allow for method chaining
}
\description{
This is inspired by Ruby's Object#tap method which allows chaining calls like:
`something.tap { |o| puts o }.method`
}
\details{
The idea is to use this function, or it's derivatives in a pipeline like:
`df %>% tap(print) %>% filter(...) %>% tap(print) %>% collect()`
}
