% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{unix2POSIXct}
\alias{unix2POSIXct}
\title{Turns a unixtimestamp into a POSIXct date}
\usage{
unix2POSIXct(time)
}
\arguments{
\item{time}{unixtimestamp to convert.}

\item{time}{class output class.}
}
\description{
Turns a unixtimestamp into a POSIXct date
}
\examples{
unix2POSIXct()
}
\keyword{time}

