% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/many_functions.R
\name{write.success}
\alias{write.success}
\title{Saves the success status of a script to a database table, to be used at the end of sourced scripts}
\usage{
write.success(con.status, table.name = "task_status")
}
\description{
Saves the success status of a script to a database table, to be used at the end of sourced scripts
}
\examples{
write.success()
}
\keyword{etl}
\keyword{pipelines,}
\keyword{status,}

