% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_functions.R
\name{cache.dated}
\alias{cache.dated}
\title{Saves an object to the cache/ dir adding the date when the object was saved}
\usage{
cache.dated(object)
}
\description{
Saves an object to the cache/ dir adding the date when the object was saved
}
\keyword{cache}

