% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/many_functions.R
\name{clean.urls.f}
\alias{clean.urls.f}
\title{Returns the cleaned url using the list of top level domains from tldextract}
\usage{
clean.urls.f(original.urls, tld = getTLD())
}
\description{
Returns the cleaned url using the list of top level domains from tldextract
}
\keyword{url}

