% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_helpers.R
\name{load.cache.dated}
\alias{load.cache.dated}
\title{load the most recent version of an object, follows the naming conventions of cache.dated()}
\usage{
load.cache.dated(object.name, cache_date = NULL, use_s3 = FALSE,
  bucket_name = NULL, force_s3_check = TRUE,
  AWS_ACCESS_KEY_ID = Sys.getenv("AWSAccessKeyId"),
  AWS_SECRET_ACCESS_KEY = Sys.getenv("AWSSecretKey"),
  AWS_DEFAULT_REGION = "eu-west-1")
}
\description{
load the most recent version of an object, follows the naming conventions of cache.dated()
}
\keyword{cache}

