% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rbindlist.valid}
\alias{rbindlist.valid}
\title{Wrapper for rbindlist that checks the stack every object in the input list that's a data.table}
\usage{
rbindlist.valid(list.data.tables.x)
}
\description{
Wrapper for rbindlist that checks the stack every object in the input list that's a data.table
}
\examples{
cache.dated()
}
\keyword{cache}

