% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_functions.R
\name{unix2POSIXct}
\alias{unix2POSIXct}
\title{Turns a unixtimestamp into a POSIXct date}
\usage{
unix2POSIXct(time)
}
\arguments{
\item{time}{unixtimestamp to convert.}

\item{time}{class output class.}
}
\description{
Turns a unixtimestamp into a POSIXct date
}
\keyword{time}

